move magic SMS mocking code into own util class
// FREEBIE
